splash screen

+++ app/src/main/java/com/example/petlover/ui/user/UserAdapter.kt
@@ -65,15 +65,17 @@ class UserAdapter (private val modelItems: ArrayList<Model>): RecyclerView.Adapt
                     Snackbar.make(view,"Done, delete success",Snackbar.LENGTH_SHORT).show()}
                 .addOnFailureListener { e -> Log.w("Error delete pet", "Error deleting document", e)
                     Snackbar.make(view,"Fail, delete fail",Snackbar.LENGTH_SHORT).show()}
-            dbImg.reference
-                .child("${model.imgRef}")
-                .delete()
-                .addOnSuccessListener {
-                    Log.d("Delete pet", "DocumentSnapshot successfully deleted!")
-                   }
-                .addOnFailureListener { e ->
-                    Log.w("ErrorDelete", "Error deleting document", e)
+            if (!model.imgRef.isNullOrEmpty()){
+                dbImg.reference
+                    .child("${model.imgRef}")
+                    .delete()
+                    .addOnSuccessListener {
+                        Log.d("Delete pet", "DocumentSnapshot successfully deleted!")
                     }
+                    .addOnFailureListener { e ->
+                        Log.w("ErrorDelete", "Error deleting document", e)
+                    }
+            }
         }
         builder.setNegativeButton("No") {dialogInterface, i ->
             Snackbar.make(view,"Ok, You cancelled the delete",Snackbar.LENGTH_SHORT).show()
